toConstant and getSymbol methods have been added to Dialect.
The toConstant default method builds a constant from a String by
inserting an underscore before each uppercase character but the
leading character, then converting the result to uppercase
The getSymbol default method returns a symbol for any NamedElement. If
that element contains a tag value named "symbol", that value is returned
otherwise a constant is built from the name of the element.