factory.vm and treebuilder.vm templates have been improved
The factory.vm template creates the XxxFactory class which contains
the registerFactories method. That method registers a default factory
for any class which has a "symbol" tag value (it must be stereotyped
with "operation").
Currently, the treebuilder generates a dummy class. It is intented to
generate a class which builds an abstract tree from an antlr concrete
syntax tree.