Parenthesis expressions have been added to the tree-builder.vm template
The visitor method for parenthesis expressions is hard coded in the
template. Its use (and the used brackets) are defined by the grammar
it self.
A specific fragment has been added to the fragments of the tree builder
template to handle exotic syntax (ternary operators, if else statement
and so on). It is not used yet.