Sequence bug has been fixed
The isOrdered method of SequenceImpl was not defined. As a consequence
a sequence.isOrdered returned the value of an instance variable and
since that variable was not initilized, the method returned false.
Now, the isOrdered method of SequenceImpl class is defined and returns
true.
The isDistinct method of SequenceImpl is defined as well