The @pre operator has been added to the concrete and abstract syntax.
In the concrete syntax the @pre operator did not work on primitive
implicite navigation (when @pre was on the first term of the navigation
which was the identifier of an attribute of the context).
This has been fixed.
The TreeBuilder has two more hard coded methods to handle @pre operator.