The JavaGenerator works fine for simple postconditions
A postcondition is translated to a set of instructions (usually a
single assignment) in ALF.
The the ALF statements are translated into Java by a dialect.
Currently, the cyclic dependencies of variables are not managed.
If a method has several postconditions, these are treated in
order of appearance which can lead to malformed code.