Add all properties to the "support" of classes in xlang-builder
A property had to be involved in an invariant to be added to
the support of its type.
In order to simplify the generation of forms, we add all properties
to the support ; as a consequence, a form can ask directly for
matching entities.
TODO : improve the JSF controllers in order to leverage the new
generated methods.