Added DialectAction
Some templates need more than one dialect.
There were two ways to achieve this goal : allow a list in the dialect
attribute or create a dialect action.
This release contains the second solution.
A very small bug in the tests of the Ete library has been fixed : the
package clause in the template class2interface.vm