Added the XmlPaths interface
This interface aims to help to customize formats of XML files.
When the reader reads an XML file, it uses XPath expressions to select
the different elements.
Depending on the XMI version, these paths can be different. The reader
will use a delegate to get these paths.
Currently, the reader implements directly the interface instead of using
a delegate.