The natural language archetype works. It is provided with a french grammar by default.
The user can customize the grammar at any time.
The usual customizations are the key words in the Lexer, the
navigation rule and the atPre rule in the parser.
The navigation can be from right to left (like in french, spanish...) or
from left to right (like in OCL, japanese...)
The atPre operator can be prefixed (like in english, japanese...) or
postfixed (like in OCL, french...)