The setPrefix method has been added to the Dialect interface in ete-lib
The default implementation does not do anything but the JavaBaseDialect
uses it to set the prefix property.
The default value of the prefix property in JavaBaseDialect is the
empty String.
The Velocity action injects the value ot the "prefix" parameter if it
is present.
A template can use this property in the generated code.