Change the support type in EnhancedInvariantImpl from Collection<MofProperty> to Collection<Step>
A Step (actually it is a full navigation) can provide the final
navigation to compute at runtime and the initial navigation as well
If the final navigation is present, we must not add the "where" clause.
The initial navigation form control must be updated when any other
initial navigation in the invariant is modified.