Create the Join class and substitute a List<Join> instance to the List<Step> instance in EteQuery class
Previously, the EteQuery was keeping a list of Steps in order to help
the SqlGenerator to build joins. But is was not simple to synchronize
the variables used in joins with those used in "where" clauses.
To fix that, it is simpler to generate Joins at the time we build the
EteQuery instance.