Fix a bug in the parsing of Preconditions in ConditionVisitor
The factory for EnhancedPreconditionImpl was not registered by the
action declaring the visitor, before the model is read. As a consequence
the Precondition the visitor received was a PreconditionImpl instance
and the related visits where not triggered.
The bug appeared in Maven process (insset-airways), not in tests since
the tests instanciated directly ConditionVisitor before they read the
model and the constructor registers the factory.
A lot of debugging code has been added to ConditionVisitor and
JavaGenerator to find the bug.