Remove the OptionalList class and add a template to the jpa module to the jsf archetype
In the JSF archetype, the OptionalList class was provided as a source
file. But the jpa module depends on that class.
As a consequence, the user who wished to use that module had to copy
the class as well.
Now the class is generated by the module. The module is autonomous.